Update license to be a mix of BSD 3-clause and CC BY 4.0
We anticipate that this will be a mixed project including both source code as well as
written and graphical content.  An open source license for software isn't optimal for
written and graphical content, while a Creative Commons license isn't optimal for
source code.  This commit changes the license for source code to BSD and all other
content to Creative Commons Attribution 4.0.

I removed capitalization for the warranty because we don't need to shout.  In
particular, formatting something in all-caps reduces readability and the warranty is
conspicuously located even without the capitalization.